pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pneumonia-detection-ai"
version = "1.0.0"
description = "Professional-grade AI system for pneumonia detection from chest X-rays using deep learning"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Shubham Singh", email = "contact@example.com"}
]
keywords = [
    "pneumonia",
    "chest-xray",
    "deep-learning",
    "tensorflow",
    "medical-imaging",
    "healthcare-ai",
    "resnet",
    "cnn",
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Healthcare Industry",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Medical/Dental Information Resources",
]
requires-python = ">=3.8"

dependencies = [
    "tensorflow>=2.13.0,<3.0.0",
    "numpy>=1.24.0,<2.0.0",
    "scipy>=1.11.0,<2.0.0",
    "scikit-learn>=1.3.0,<2.0.0",
    "pandas>=2.0.0,<3.0.0",
    "Pillow>=10.0.0,<11.0.0",
    "matplotlib>=3.7.0,<4.0.0",
    "seaborn>=0.12.0,<1.0.0",
    "plotly>=5.14.0,<6.0.0",
    "streamlit>=1.28.0,<2.0.0",
    "streamlit_extras>=0.3.0,<1.0.0",
    "streamlit_option_menu>=0.3.0,<1.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "pydantic>=2.0.0,<3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.9.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
    "isort>=5.12.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/pneumonia-detection-ai"
Documentation = "https://github.com/yourusername/pneumonia-detection-ai/wiki"
Repository = "https://github.com/yourusername/pneumonia-detection-ai"
Issues = "https://github.com/yourusername/pneumonia-detection-ai/issues"

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py10', 'py11']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
